trans opt collection indexed by coverage & stored pointers, rather than obj.
fixed mem leak in manager


git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@333 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/SquareMatrix.h
@@ -45,7 +45,7 @@ public:
 	{
 		free(m_array);
 	}
-	inline float GetSize() const
+	inline size_t GetSize() const
 	{
 		return m_size;
 	}
